Fork qng as qngng
