+++ MANIFEST.in
@@ -1,1 +1,1 @@
-recursive-include qng/data *
+recursive-include qngng/data *
