--- qng/__init__.py
@@ -1,2 +1,0 @@
-__version__ = '0.1.3'
-__description__ = 'The Queb name generator'
