+++ qngng/__init__.py
@@ -1,0 +1,2 @@
+__version__ = '0.2'
+__description__ = 'The Queb name generator: next generation'
