+++ setup.cfg
@@ -1,13 +1,13 @@
 [metadata]
-name = qng
-version = attr: qng.__version__
-description = The Queb name generator
+name = qngng
+version = attr: qngng.__version__
+description = The Queb name generator: next generation
 long_description = file: README.rst
-url = https://github.com/abusque/qng
-author = Antoine Busque
-author_email = antoinebusque@gmail.com
+url = https://github.com/eepp/qngng
+author = Philippe Proulx, Antoine Busque
+author_email = eeppeliteloop@gmail.com
 license = MIT
-keywords = name, generator, quebec
+keywords = name, generator, quebec, queb, uda
 classifiers =
     License :: OSI Approved :: MIT License
     Programming Language :: Python
@@ -20,5 +20,5 @@ include_package_data = True
 
 [options.entry_points]
 console_scripts =
-    qng = qng.cli_qng:main
-    qng-uda = qng.cli_qng_uda:main
+    qngng = qngng.cli_qngng:main
+    qngng-uda = qngng.cli_qngng_uda:main
